backend_rds/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.61
botocore==1.31.61
cffi==1.16.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
exceptiongroup==1.1.3
fastapi==0.103.2
fonttools==4.43.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
jwt==1.3.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
psycopg2==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
uvicorn==0.23.2
